fix(checker): split computed property name helpers
